Change the default values of database settings
We chose 100 for the max connections, as this is the default value for a
single postgres server. This should run fine, if you only run a single
controller.
The minimum amount of open connections at any given time is set to 10 as
a sane default.
This scales up to max connections if the load increases (requests are
waiting for connections from our connection pool)