Add reload flag
Added a flag `--reload`. If that flag is set, the controller is not intended to start normally, instead it tries to find another controller process and sends a SIGHUP to it.
The idea is to be able to trigger a partial reload of the controller, to adjust to changed configuration values.