No longer execute module requested actions inside event handlers
Instead return the requested actions to the caller. This allows
the caller to control when these actions were taken. This solves the
problem of modules sending websocket messages to the frontend
before it might be properly initialized

E.g. ParticipantJoined event used to send a websocket message
after sending an event to every module, but the modules messages
were sent before the original ParticipantJoined event, leading
references to a participant the frontend has no knowledge of (yet).

Same goes for the Joined event, websocket messages of modules
would be sent before the initial JoinSuccess message.