Introduce timestamp for all messages.
The origin of the timestamps is either when the websocket message enters
our system or in the case of events originating in our system the time
they occur the first time.

E.g. User Joined messages will have the timestamp of when the join
message arrived at our websocket. User Left will have the timestamp of
when the event loop of this user ended.