improve CI container build
Start a container dry build without pushing the image to check the container build
still works on merge requests. As we have no access to the container
registry on foreign repositories, we only run this new job in MRs. As
any limitation to a merge requests, excludes all other jobs from the
merge pipeline, this commit also adds a few rules to all jobs.