Add the moderator mute feature to the media module
Adds the `ModeratorMute` feature to the janus-media module, that allows moderators to manage the mute status of other participants.

The moderator creates a list of participants that shall get muted and sends it to the controller. The controller itself does not
change the mute status of participants but rather sends a `RequestMute` to the affected participant/s.

The `RequestMute` message contains a flag to differ between a forced or non-forced mute. The non-forced mute should just notify
the participant to mute the microphone. The forced variant should automatically mute the microphone and notify the participant about
the mute.

Related: k3k-web-frontend#148
Closes: #92