db-storage: use a consistent id schema across the whole db model
Each 'entity' now has a `id` field which is of the type UUID with a
default value of `gen_random_uuid` and a `id_serial` field which is
of the type BIGSERIAL. All references must be to the `id` field, as
the uuid is preffered for our REST APIs as it doesn't disclose how
many of each entity exists. Just in case we want to switch back to BIGSERIAL
ids we keep a `id_serial` field as we may encounter problems with the UUID type.

Also to prepare for support for multiple OIDC providers we now store
the `oidc_issuer`, which is extracted from the `iss` field in the
JWT tokens, in the `users` table. To prevent collisions between `sub`s
from different OIDC providers we now generate our own id for each user
(just like for every entity as described above) and store the `sub` from
the JWT token in the new `oidc_sub` field which is now a VARCHAR(255)
as specified in https://openid.net/specs/openid-connect-core-1_0.html#IDToken

Note that, with the rather big migration, the order of fields in the models
has changed according to the modification of the type. This is a drawback
of using diesel's automatic schema generation as it requires the model's
fields to have the same order as the columns in the database.
This can be changed back in the future if we decide to merge all migrations
into one before deploying the software anywhere where we dont have any control
over it anymore. (Even without breaking any testing/staging databases as they
can be migrated with the old migrations. Out of order columns are not a problem
as diesel will never do a `SELECT * FROM ...` but always specify all fields.)

Since discovering that tracing is able to automatically emit a error
message if a function returns `Result::Err` i've removed most explicit
error logs and result 'inspections' from the db functions, and
added the `err` parameter to the tracing instrument macro.