controller: lock room mutex while adding attributes and use defaults when building participants
Also set all redis attributes when joining, not on runner create.

If the last participant left a room all attributes and data related to the room is deleted.
Participant not yet joined or in the process of joining will have their data removed while joining
leaving them invisible as other runner will not be able to fetch their attributes, skipping them.

While setting/removing all attributes only while having the redlock might suffice,
this commit also adds defaults for all attributes (at least the ones used by the runner)
so that even if we fail to fetch some attributes the participant will be displayed in
some form making them visible to the frontend.