kustos: add policies_builder module for more ergonomic policy building
Currently when setting/removing permissions using kustos we've had to
define all access methods and endpoints for each resource.

Using the PoliciesBuildere we can build a set of policies for different
subjects (users, groups, roles) and then add/remove them all at once.

By leveraging extensions traits (like the RoomsPoliciesBuilderExt) which are
implemented on the PoliciesBuilder we can define all resources (and the access
to them) in one place.