remove ci builds on feature branches
At the moment we're running most of the pipeline on feature branches and
on MRs. This leads to twice the load on our runners when a MR is open
and updated since all the tasks are run twice, once for the MR and once
for the feature branch. This MR removes the build for the feature
branch. Thus if anyone wants to get a build on a feature branch one
would to run the pipeline manually.