controller: Replace tungstenite with actix-web-actors
This switches to actix websocket actors because the old attempt
to use tungtenite with web::Payload and a Stream->AsyncRead adapter
proved to be quite error prone and had stuck connections.

An error fix in the newest actix-web release fixed the stuck
connection error but then sends a 500 Internal Server error if the
web::Payload object is dropped (apparently) prematurely.

Faster shutdowns can now be observed as actix-web is now able
to immediately disconnect any open connections. Before it
(almost always) ran into the configured max wait timeout for the shutdown.

Co-authored-by: Rudi <r.floren@heinlein-video.de>