Remove matching_fn from the rbac manager from casbin
Using this has bad scaling performance in combination with the
auto_loader.

We rely on the auto loader to allow the current HA setup. The casbin
policy model is fetched every 10 s from the database.
During loading the rbac links are completely flushed and reculculated.
This call has many nested loops that call the matching fn at the lowest
level. This function creates regexs on runtime during that call which
will result in the rbac link recalculation to take longer than 10 s in the
worst case with less than 1500 users. :(