Move introspection endpoint to another module
Currently this module is used in the internal http server
As it is uncertain if this is used in the future we move this here as
this is not considered part of the api but an internal api.

The internal module in the api module is reserved for public accessible
but internally used API endpoints (e.g. special endpoints for the
current frontend)