Add internal endpoint for deleting a room
This endpoint is used by the frontend to delete a room and all related
events which are not hardly linked.

The normal room delete endpoints returns a 409 when there are still
referencing events present. This is needed to avoid accidential event
deletion by third party REST API clients.
The frontend wanted a transactionless way to delete the room with the
linked events.