controller/signaling: add waiting room to moderation module
This is achieved via a combination of logic inside the websocket runner
and moderation module. The actual state of the websocket is saved inside the
runner where it differs between being inside the waiting room and the actual room.

There now 2 sets of participant-ids to track when accounting for thhe  rooms lifetime.
The participant set which existed before and now the waiting-room set.

Participants inside the waiting room will have the control-module atttributes set in redis.
This is needed that their information like display_name can be displayed for moderators
which join after the participant joined the waiting room.

Since all attributes are bound to the rooms lifetime the rooms must only be cleaned up
if both the participant- AND waiting-room set are empty.

All the user facing logic run via the `moderaiton` module meaning websocket messages
are all contained inside the `moderation` namespace. The only exception is a new
control-rmq message, which is emitted by the `moderation` module via a new crate-private
member function on the `ModuleContext` `rabbitmq_publish_control`, which indicates
a runner that the participant was accepted into the room and should now join the room
and leave the waiting room.