Add .dockerignore file
Docker copies all files from the working directory to the builder
context (the one that builds the images, not defined builder images).
When you have a large target dir there are a couple of GB copied
everytime you build a container. Adding this avoids this, shaving a
couple of seceonds of the container build time.