controller: remove GET /users and add settings to disable /users/find and set default user language
The `GET /users` was not thoroughly thought out when implemented, isn't used anywhere
and relies on potentially volatile information stored in casbin. Thats why this
commit removes it for now. Can be reimplemented corrently once needed.

Option to disable the /users/find endpoint is needed as tenancy isn't implemented
yet and to limit discoverability of other users this endpoint has to be disabled.

Language default for users is needed to be configurable since not all
deployments want en-US to be the language default.