mails: Add initial invite notification sending
This adds the mail-worker-proto crate, an interface definition for
sending mail sending tasks to another worker.
Currently this is done using rabbitmq. In the future we can support the
usage of a in binary mail sender. For now this is a worker run as a
different process.
Currently this supports only the invite mail for registered users.