controller/signaling: add user phone mapping
Adds the functionality to map a phone number to a users display name.
This feature can be enabled/disabled in the call_in settings.

A users phone number cannot be configured manually. It gets set on login
if the id_token contains the `phone_number` field.

The phone/user mapping fails if more than one participant have the same
number configured.

Changed the call-in phone number to the international format when possible.