Add the spacedeck module
Allows participants to collaboratively edit a whiteboard.

Moderators can initiate the rooms 'space' with the `Initialize` message. All participants will receive the access url for the space.

Joining participants receive the rooms space-state via the module data in the `join_success` message.

Moderators can generate a PDF document from the current content of the space via the `GeneratePdf` message. The URL to access the PDF is received by every participant in the room.