janus-media: configure media relay on mute
Whenever a participants mutes one of his media streams (e.g. video or audio)
we get a message with the updated MediaSessionState. Historically we've relied
on the client to properly configure their local streams to stop sending data
from a muted media source and the backend only propagated their mute-state
to other participants.

We've since learned that browsers do in fact not stop sending their data
but instead transmit empty/silent RTP packets (for audio at least).
The webrtc-spec requires to stop sending audio packets when disabled or
one black frame per second for video.
Sadly that spec doesn't seem to be completely followed.

That by itself is not a problem since privacy is ensured by sending empty data
but it's a lot of packets to containing no information. Also it created a lot
of unnecessary load on janus and the frontend processing that many empty/silent
packets.

Now once a new MediaSessionState is received, it's used to stop the relay of
muted media streams on the publisher side. That means all packets containing
no information will be discarded instead of forwarded to the subscribers.