Add source and type information to error response
Normally, in a conference the client would issue a lot of subscribe and configure requests in
parallel and would not be able to known which requests caused an error.

With adding the source and type to the error a FE can map the error, as normally the client would only send one subscribe/configure per target x type
combination.