controller: Add S3 MinIO storage interface
Adds an S3 storage interface, allowing the controller and modules to save assets in long-term storage.

Each saved asset has a database entry with some metadata attached. Assets can be bound to a room.

The S3 access key for an asset is its ID in the database entry.

Adds REST API endpoints to get, list and delete assets.