controller: add service endpoints with special authentication
To support services like call-in and recording the controller needs
to be able to authenticate other parts of the infrastructure.

Our user authentication and authorization relies on access-tokens
acquired via traditional oidc/oauth2 flows.

To authorize services we also use access-tokens. The administrator
creates a new oidc-client and service-account for a service in keycloak.
The service then can fetch an access-token using the 'client_credentials'
flow.

If configured correctly the access-token will contain roles assigned to
the service account. Using these roles the controller can decide if
the caller of the API has permissions to access it.

All endpoints inside the `/services/*` path require the caller to have
certain roles set in their access-token.

For example: With this commit the start-endpoint for the call-in service
changed from `POST /rooms/sip/start` to `POST /services/call_in/start`
and checks if the caller has the role "opentalk-call-in".