Add hidden legal votes
Added the `hidden` flag to the legal vote user parameters. When a legal
vote is configured to be hidden, the outgoing `Updated` and `Stopped`
message won't contain the participants vote choices.

Made the user information in the `Vote` event in the legal vote protocol
optional. The user information is omitted when the vote is configured to
be hidden.