legal-vote: add PDF capabilities
Adds the functionality to render finished vote protocols as a PDF
document. The documents get saved in the S3 storage and the asset ID is
sent to the responsible moderator.

When the `create_pdf` flag is set at the start of the vote, the protocol
PDF is automatically created and saved once a vote has ended.
In case of an automatic stop, the PDF asset ID is sent to the vote
initiator.

Moderators can manually generate PDF documents of passed votes via the
`incoming::GeneratePdf` message. This also includes failed or canceled
votes. The issuing moderator will get a `PdfAsset` response containing
the `AssetId` of the created document.