remove half baked support for multiple oidc provider
We once planned to support multiple OIDC provider (Keycloak)
as can be seen on some other sites where you can login via a
bunch of other sites (e.g. Github, Google, Microsoft etc...).
The current implementation is error prone and needs a complete rework.
To simplify the implementation of the multi-tenancy we've decided to
completely remove this feature and implement it properly later if we still want it.

Currently we've been storing the complete issuer-url in every user and group.
This will break if the deployment changes the url or the name of the realm.
To properly implement this we should have the OIDC configs referenced instead of copied into every row.