timer: add `style` field and improve API
Adds the `style` field to the `start` & `started` messages to let
clients tag a timer with a custom style. The style is included in the
`started` message so that a frontend application can handle different
styles accordingly.

Adds the `kind` field to the `start` & `started` messages to make the
configuration of the timer more obvious. Previously, the timer module
deduced the timer kind (stopwatch/countdown) from the given duration.