Add multi tenant support
Multiple tenant is implemented by attaching a `tenant_id` to all resources.

The controller now requires all JWT-claims to contain an `tenant_id`.
It can be any string identifying a single tenant. When the controller
encounters a new `tenant_id` it will create a new tenant (if it doesn't
exist) and add the user to the tenant.

Users from the keycloak are now identified using the `sub` and `tenant_id`
claim instead of just `sub`. This enables users to switch tenants without
opentalk needing to migrate user specific resources.

All queries that allow for user discovery and interaction have been modified
to also check the `tenant_id` for equality to avoid information crossing tenants.