controller: add `participant_limit` quota
Adds support for a `"participant_limit"` quota. The quota will be enforced when a participant
attempts to join a room via the `control::incoming::Join` message. When the configured
`participant_limit` of a room is reached, a `control::outgoing::JoinFailed` message is sent
as a response, containing the reason `participant_limit_reached`.

closes: #332