kustos: add `macros` feature on `tokio` dependency
This feature is required for calling the `tokio::select!` macro.
The call is not behind a feature gate, so it should be available
unconditionally. Apparently the issue did not surface until now, because
wherever we have a dependency on the `kustos` crate, we also have a
dependency on `tokio` with the `macros` feature enabled.